refactor: extract community reports hook
