+++ frontend/src/features/participant/community/CommunityPage.jsx
@@ -37,7 +37,6 @@ import {
 import {
   getCurrentCommunityUserId,
   getFollowAuthorKey,
-  getPostAuthorId,
   isAuthorCurrentUser,
   isAuthorFollowed,
   isPostOwnedByCurrentUser,
@@ -50,10 +49,6 @@ import {
   saveStoredCommunityUserProfile,
   saveStoredFollowedAuthors,
 } from './services/communityStorageService';
-import {
-  reportCommunityPost,
-} from './services/communityService';
-import { COMMUNITY_POST_STATUS } from './communityModels';
 import {
   COMMUNITY_GUIDELINES_VERSION,
   getAcceptedGuidelinesVersion,
@@ -73,6 +68,7 @@ import FeedTabs from './components/FeedTabs';
 import ReportPostModal from './components/ReportPostModal';
 import useCommunityComments from './hooks/useCommunityComments';
 import useCommunityPosts from './hooks/useCommunityPosts';
+import useCommunityReports from './hooks/useCommunityReports';
 import './styles/community.css';
 
 export default function CommunityPage({
@@ -99,10 +95,6 @@ export default function CommunityPage({
   const [communityUserProfile, setCommunityUserProfile] = useState(getInitialCommunityUserProfile);
   const [communityPreferences, setCommunityPreferences] = useState(getInitialCommunityPreferences);
   const [profileSuccessMessage, setProfileSuccessMessage] = useState('');
-  const [reportFeedbackByPostId, setReportFeedbackByPostId] = useState({});
-  const [confirmingReportPostId, setConfirmingReportPostId] = useState(null);
-  const [selectedReportReason, setSelectedReportReason] = useState('');
-  const [reportReasonError, setReportReasonError] = useState('');
   const [confirmingDeletePostId, setConfirmingDeletePostId] = useState(null);
   const [editingPostId, setEditingPostId] = useState(null);
   const [editPostText, setEditPostText] = useState('');
@@ -115,16 +107,6 @@ export default function CommunityPage({
   const [isCommunityStreakAtRisk, setIsCommunityStreakAtRisk] = useState(
     () => isStreakAtRiskForDate(initialStreakState.lastActivityDate),
   );
-
-  useEffect(() => {
-    if (!confirmingReportPostId) return;
-
-    setSelectedReportReason('');
-    setReportReasonError('');
-    window.setTimeout(() => {
-      reportModalRef.current?.focus();
-    }, 0);
-  }, [confirmingReportPostId]);
 
   useEffect(() => {
     if (!confirmingDeletePostId) return;
@@ -281,6 +263,29 @@ export default function CommunityPage({
     registerCommunityActivity,
     setPosts,
   });
+  const {
+    reportFeedbackByPostId,
+    confirmingReportPostId,
+    selectedReportReason,
+    reportReasonError,
+    handleReportPostRequest,
+    handleCancelReportPost,
+    handleReportReasonChange,
+    handleReportSubmit,
+  } = useCommunityReports({
+    communityDisplayName,
+    localUserId,
+    posts,
+    setPosts,
+  });
+  useEffect(() => {
+    if (!confirmingReportPostId) return;
+
+    window.setTimeout(() => {
+      reportModalRef.current?.focus();
+    }, 0);
+  }, [confirmingReportPostId]);
+
   useEffect(() => {
     if (!pendingCommentDeletion) return;
 
@@ -365,42 +370,6 @@ export default function CommunityPage({
     ));
   };
 
-  const handleReportPostRequest = (postId) => {
-    const postToReport = posts.find((post) => post.id === postId);
-
-    if (isPostOwnedByCurrentUser(postToReport, localUserId, communityDisplayName || 'Current User')) {
-      setConfirmingReportPostId(null);
-      return;
-    }
-
-    if (isPostReportedByUser(postToReport, localUserId)) {
-      setConfirmingReportPostId(null);
-      setReportFeedbackByPostId((currentFeedback) => ({
-        ...currentFeedback,
-        [postId]: {
-          type: 'error',
-          message: 'You already reported this post.',
-        },
-      }));
-      return;
-    }
-
-    setReportFeedbackByPostId((currentFeedback) => {
-      if (!currentFeedback[postId]) return currentFeedback;
-
-      const nextFeedback = { ...currentFeedback };
-      delete nextFeedback[postId];
-      return nextFeedback;
-    });
-    setConfirmingReportPostId(postId);
-  };
-
-  const handleCancelReportPost = () => {
-    setConfirmingReportPostId(null);
-    setSelectedReportReason('');
-    setReportReasonError('');
-  };
-
   const handleReportModalBackdropClick = (event) => {
     if (event.target === event.currentTarget) {
       handleCancelReportPost();
@@ -411,97 +380,6 @@ export default function CommunityPage({
     if (event.key === 'Escape') {
       handleCancelReportPost();
     }
-  };
-
-  const handleReportReasonChange = (reason) => {
-    setSelectedReportReason(reason);
-    if (reportReasonError) setReportReasonError('');
-  };
-
-  const handleReportSubmit = (event) => {
-    event.preventDefault();
-
-    if (!confirmingReportPostId) return;
-
-    if (!selectedReportReason) {
-      setReportReasonError('Please choose a report reason before submitting.');
-      return;
-    }
-
-    handleConfirmReportPost(confirmingReportPostId, selectedReportReason);
-  };
-
-  const handleConfirmReportPost = async (postId, reason) => {
-    const postToReport = posts.find((post) => post.id === postId);
-
-    if (!postToReport) return;
-
-    if (isPostOwnedByCurrentUser(postToReport, localUserId, communityDisplayName || 'Current User')) {
-      setConfirmingReportPostId(null);
-      setSelectedReportReason('');
-      setReportReasonError('');
-      return;
-    }
-
-    if (isPostReportedByUser(postToReport, localUserId)) {
-      setConfirmingReportPostId(null);
-      setReportFeedbackByPostId((currentFeedback) => ({
-        ...currentFeedback,
-        [postId]: {
-          type: 'error',
-          message: 'You already reported this post.',
-        },
-      }));
-      return;
-    }
-
-    const createdAt = new Date();
-    const reportRecord = {
-      id: `community-report-${postId}-${localUserId}-${createdAt.getTime()}`,
-      postId,
-      reporterUserId: localUserId,
-      postOwnerId: getPostAuthorId(postToReport) || postToReport.authorDisplayName || postToReport.author || null,
-      reason,
-      createdAt: createdAt.toISOString(),
-    };
-
-    try {
-      await reportCommunityPost(postId, reportRecord);
-    } catch {
-      // Keep the local placeholder flow responsive; Firebase reporting will handle failures later.
-    }
-
-    setPosts((currentPosts) => currentPosts.map((post) => {
-      if (post.id !== postId) return post;
-
-      const currentReportedBy = Array.isArray(post.reportedBy) ? post.reportedBy : [];
-      if (currentReportedBy.includes(localUserId)) return post;
-
-      const reportsCount = Number(post.reportsCount);
-      const nextReportsCount = Number.isFinite(reportsCount) && reportsCount >= 0
-        ? reportsCount + 1
-        : 1;
-
-      return {
-        ...post,
-        reportsCount: nextReportsCount,
-        reportedBy: [...currentReportedBy, localUserId],
-        reports: [...(Array.isArray(post.reports) ? post.reports : []), reportRecord],
-        status: post.status === COMMUNITY_POST_STATUS.active
-          ? COMMUNITY_POST_STATUS.reported
-          : post.status ?? COMMUNITY_POST_STATUS.reported,
-      };
-    }));
-    setConfirmingReportPostId(null);
-    setSelectedReportReason('');
-    setReportReasonError('');
-    setReportFeedbackByPostId((currentFeedback) => ({
-      ...currentFeedback,
-      [postId]: {
-        type: 'success',
-        message: 'Thanks, your report was saved locally for review.',
-      },
-    }));
   };
 
   const handleDeleteCommentModalBackdropClick = (event) => {
